Refactor logging framework to modular API
- Introduced a modular API structure by creating separate modules for the logging library, including `api`, `loggers`, and `constants`.
- Added new files for `api.lua`, `loggers.lua`, and `constants.lua` to enhance organization and maintainability.
- Updated the main logger module to utilize the new API structure, improving clarity and separation of concerns.
- Enhanced the rockspec file to include new modules, ensuring proper loading and dependency management.
- Refactored existing code to align with the new modular design, promoting better extensibility and future enhancements.
