+++ lua/lual/async/config.lua
@@ -1,6 +1,8 @@
 --- Async Configuration Handler
 -- This module handles the 'async' configuration key
 
+-- Note: For direct execution with 'lua', use require("lua.lual.*")
+-- For LuaRocks installed modules or busted tests, use require("lual.*")
 local async_writer = require("lual.async")
 
 local M = {}
